Har rettet ting jeg ikke har forstand på. Virker forhåbentligt

+++ warehouse.h
@@ -9,14 +9,14 @@ struct node {
     int neighbour_count;
     edge** successors; //array af pointere
     edge** predecessors; //array af pointere
-    int* g;
-    int* h;
-    int* rhs;
+    float* g;
+    float* h;
+    float* rhs;
 };
 
 struct edge {
     node* source;
-    double cost;
+    float cost;
     node* dest;
 };
